Add autosubscribe argument to fetchFullAccount.
Whenever fetchFullAccount is called, the server auto-subscribes the client to receive notifications when data about the account changes. This can be problematic when fetching data for over 100 accounts. On the server side, an assertion fails and causes an exception to be thrown. On the client side, this error is ignored and the client continues to repeat the the get_full_account command about every 5 seconds. This can cause significant lag in the client as network an processing cycles are continually wasted on operations that will never succeed. Thie commit allows a client to decide if it should attempt to be autoregistered or not.